pyproject.toml
[project]
name = "orkestra"
version = "0.1.4"
description = "For orchestration using AWS Step Functions"
authors = [
    {name = "Stephan Fitzpatrick", email = "knowsuchagency@gmail.com"},
]
dependencies = [
    "aws-cdk.core",
    "aws-cdk.aws_events",
    "aws-cdk.aws_events_targets",
    "aws-cdk.aws_lambda_python",
    "aws-cdk.aws_stepfunctions_tasks",
]
requires-python = ">=3.8"
dynamic = ["classifiers"]
license = {text = "MIT"}
keywords = ["orchestration", "aws", "cloud"]


[project.urls]
homepage = "https://github.com/knowsuchagency/composer"
Repository = "https://github.com/knowsuchagency/composer"

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool.pdm.scripts]
_.env_file = ".env"

synth = "cdk synth"
deploy = "cdk deploy --verbose"
destroy = "cdk destroy"

test-publish = "twine upload --repository testpypi dist/*"

publish.shell = """

twine upload dist/*
"""

pub.shell = """
set -x
pdm build
pdm run test-publish
pdm run publish
"""

[tool.pdm.dev-dependencies]
dev = [
    "loguru~=0.5",
]
